+++ packages/typegen/src/util/derived.ts
@@ -5,7 +5,7 @@
 import { TypeDef, TypeDefInfo } from '@polkadot/types/create/types';
 import { Constructor, Registry } from '@polkadot/types/types';
 
-import { ClassOf, ClassOfUnsafe, getTypeDef } from '@polkadot/types/create';
+import { ClassOfUnsafe, getTypeDef } from '@polkadot/types/create';
 import AbstractInt from '@polkadot/types/codec/AbstractInt';
 import Compact from '@polkadot/types/codec/Compact';
 import Enum from '@polkadot/types/codec/Enum';
@@ -118,7 +118,7 @@ export function getSimilarTypes (definitions: object, registry: Registry, _type:
     possibleTypes.push('Address', 'AccountId', 'AccountIndex', 'string', 'Uint8Array');
   } else if (isChildClass(GenericAccountId, Clazz)) {
     possibleTypes.push('string', 'Uint8Array');
-  } else if (isChildClass(ClassOf(registry, 'bool'), Clazz)) {
+  } else if (isChildClass(registry.createClass('bool'), Clazz)) {
     possibleTypes.push('boolean', 'Uint8Array');
   } else if (isChildClass(Null, Clazz)) {
     possibleTypes.push('null');