+++ packages/types/src/codec/utils/index.ts
@@ -6,5 +6,7 @@ export { default as compareArray } from './compareArray';
 export { default as compareMap } from './compareMap';
 export { default as compareSet } from './compareSet';
 export { default as decodeU8a } from './decodeU8a';
+export { default as mapToTypeMap } from './mapToTypeMap';
+export { default as typeToConstructor } from './typeToConstructor';
+
 export * from './encodeTypes';
-export * from './mapToTypeMap';