+++ packages/types/src/create/types.ts
@@ -23,6 +23,7 @@ export enum TypeDefInfo {
   Tuple,
   Vec,
   VecFixed,
+  HashMap,
   Int,
   UInt,
   // anything not fully supported (keep this as the last entry)